wa_history/requirements.txt
ibmcloudant>=0.7.1
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv
